[FE] FEAT: todo  style 로직 분리
